requirements.txt
python-docx==1.1.2
pypdf==5.1.0
sentence-transformers==3.3.1
chromadb==0.5.23
numpy==1.26.4
pandas==2.2.3
openpyxl==3.1.5
tqdm==4.67.1
rapidfuzz==3.10.1
python-dotenv==1.0.1
openai==1.45.0